[WIP] refact js files
