+++ src/js/main/controls/downloadButton.js
@@ -1,0 +1,6 @@
+import prepareDownload from "../functions/prepareDownload";
+
+const downloadButton = document.getElementById("downloadButton");
+downloadButton.onclick = prepareDownload;
+
+export default downloadButton;
