+++ src/js/main/controls/loadFileButton.js
@@ -1,0 +1,6 @@
+import openFile from "../functions/openFile";
+
+const loadFileButton = document.getElementById("loadfileButton");
+loadFileButton.onclick = openFile;
+
+export default loadFileButton;
