+++ src/js/main/controls/modeButton.js
@@ -1,0 +1,6 @@
+import changeMode from "../functions/changeMode";
+
+const modeButton = document.getElementById("modeButton");
+modeButton.onclick = changeMode;
+
+export default modeButton;
