+++ src/js/main/controls/viewButton.js
@@ -1,0 +1,6 @@
+import changeView from "../functions/changeView";
+
+const viewButton = document.getElementById("viewButton");
+viewButton.onclick = changeView;
+
+export default viewButton;
