+++ src/js/main/controls/worldNameForm.js
@@ -1,0 +1,3 @@
+const worldNameForm = document.getElementById("worldnameform");
+
+export default worldNameForm;
