+++ src/js/main/functions/init.js
@@ -1,0 +1,8 @@
+import VoxelWorld from "../classes/VoxelWorld";
+import voxelFaces from "../constants/voxelFaces";
+
+function init() {
+  VoxelWorld.faces = voxelFaces;
+}
+
+export default init;
