+++ src/js/main/index.js
@@ -1,5 +1,5 @@
 import * as THREE from "three";
-import TextureImage from "../../../assets/img/texture.png";
+
 import changeMode from "./functions/changeMode";
 import cameracontrols, { camera, canvas } from "./functions/cameracontrols";
 import {
@@ -13,18 +13,16 @@ import {
 import player from "./controls/player";
 import speedSlider from "./controls/speedSlider";
 import worlds from "./controls/worlds";
+import modeButton from "./controls/modeButton";
+import currentObjectId from "./controls/currentObjectId";
+import init from "./functions/init";
+import createWorld from "./functions/createWorld";
+
+init();
 
 const c = 5;
 
 // Input Control
-const modeButton = document.getElementById("modeButton");
-modeButton.onclick = changeMode;
-const viewButton = document.getElementById("viewButton");
-viewButton.onclick = changeView;
-const downloadButton = document.getElementById("downloadButton");
-downloadButton.onclick = prepareDownload;
-const loadFileButton = document.getElementById("loadfileButton");
-loadFileButton.onclick = openFile;
 
 const mytimeElement = document.getElementById("mytime");
 
@@ -54,140 +52,12 @@ objSubmit.onclick = setWorldVelocity;
 
 var editing = true;
 var usingRelativity = 0;
-var currentObjectId;
 
 var worldNum = 0;
 var controllock = false;
 
 function lockControl() {
   controllock = true;
-}
-
-function changeView() {
-  viewButton.blur();
-  if (usingRelativity == 0) {
-    viewButton.value = "Weak Relativity";
-    viewButton.style.background = "orange";
-    usingRelativity = 1;
-  } else if (usingRelativity == 1) {
-    viewButton.value = "Strong Relativity";
-    viewButton.style.background = "red";
-    usingRelativity = 2;
-  } else {
-    viewButton.value = "Relativity OFF";
-    viewButton.style.background = "white";
-    usingRelativity = 0;
-  }
-}
-
-function doLengthContraction(x, y, z, ux, uy, uz) {
-  const beta = Math.sqrt(ux * ux + uy * uy + uz * uz) / c;
-  if (beta == 0) {
-    return [x, y, z];
-  }
-  const gamma = 1 / Math.sqrt(1 - beta * beta);
-  const factor = 1 / gamma - 1;
-  const px = ux / Math.sqrt(ux * ux + uy * uy + uz * uz);
-  const py = uy / Math.sqrt(ux * ux + uy * uy + uz * uz);
-  const pz = uz / Math.sqrt(ux * ux + uy * uy + uz * uz);
-  const xprime =
-    (1 + factor * px * px) * x + factor * px * py * y + factor * px * pz * z;
-  const yprime =
-    factor * px * py * x + (1 + factor * py * py) * y + factor * py * pz * z;
-  const zprime =
-    factor * px * pz * x + factor * py * pz * y + (1 + factor * pz * pz) * z;
-  return [xprime, yprime, zprime];
-}
-
-function applyRelativity1(x0, y0, z0, vx, vy, vz, ux, uy, uz, tb) {
-  const v = Math.sqrt(vx * vx + vy * vy + vz * vz);
-  const beta = v / c;
-  const gamma = 1 / Math.sqrt(1 - beta * beta);
-  const betau = Math.sqrt(ux * ux + uy * uy + uz * uz) / c;
-  const gammau = 1 / Math.sqrt(1 - betau * betau);
-  const ta =
-    (gamma * c * c * tb + vx * x0 + vy * y0 + vz * z0) /
-    (c * c - ux * vx - uy * vy - uz * vz);
-
-  const xprime = x0 + ux * ta;
-  const yprime = y0 + uy * ta;
-  const zprime = z0 + uz * ta;
-
-  const tobj = ta / gammau - (gammau * (ux * x0 + uy * y0 + uz * z0)) / (c * c);
-
-  if (v == 0) {
-    return [xprime, yprime, zprime, tobj];
-  }
-
-  const factor = gamma - 1;
-  const px = vx / v;
-  const py = vy / v;
-  const pz = vz / v;
-  const xb =
-    (1 + factor * px * px) * xprime +
-    factor * px * py * yprime +
-    factor * px * pz * zprime +
-    gamma * v * (gamma * tb - ta) * px;
-  const yb =
-    factor * px * py * xprime +
-    (1 + factor * py * py) * yprime +
-    factor * py * pz * zprime +
-    gamma * v * (gamma * tb - ta) * py;
-  const zb =
-    factor * px * pz * xprime +
-    factor * py * pz * yprime +
-    (1 + factor * pz * pz) * zprime +
-    gamma * v * (gamma * tb - ta) * pz;
-  return [xb, yb, zb, tobj];
-}
-
-function applyRelativity2(x0, y0, z0, vx, vy, vz, ux, uy, uz, tb) {
-  const v = Math.sqrt(vx * vx + vy * vy + vz * vz);
-  const beta = v / c;
-  const usquare = ux * ux + uy * uy + uz * uz;
-  const ud = ux * x0 + uy * y0 + uz * z0;
-  const gamma = 1 / Math.sqrt(1 - beta * beta);
-  const betau = Math.sqrt(ux * ux + uy * uy + uz * uz) / c;
-  const gammau = 1 / Math.sqrt(1 - betau * betau);
-  const dsquare = x0 * x0 + y0 * y0 + z0 * z0;
-  const paren1 = ud + gamma * c * c * tb;
-  const ta =
-    (paren1 -
-      Math.sqrt(
-        paren1 * paren1 +
-          (c * c - usquare) * (dsquare - gamma * gamma * c * c * tb * tb)
-      )) /
-    (c * c - usquare);
-  const xprime = x0 + ux * ta;
-  const yprime = y0 + uy * ta;
-  const zprime = z0 + uz * ta;
-
-  const tobj = ta / gammau - (gammau * ud) / (c * c);
-
-  if (v == 0) {
-    return [xprime, yprime, zprime, tobj];
-  }
-
-  const factor = gamma - 1;
-  const px = vx / v;
-  const py = vy / v;
-  const pz = vz / v;
-  const xb =
-    (1 + factor * px * px) * xprime +
-    factor * px * py * yprime +
-    factor * px * pz * zprime +
-    gamma * v * (gamma * tb - ta) * px;
-  const yb =
-    factor * px * py * xprime +
-    (1 + factor * py * py) * yprime +
-    factor * py * pz * zprime +
-    gamma * v * (gamma * tb - ta) * py;
-  const zb =
-    factor * px * pz * xprime +
-    factor * py * pz * yprime +
-    (1 + factor * pz * pz) * zprime +
-    gamma * v * (gamma * tb - ta) * pz;
-  return [xb, yb, zb, tobj];
 }
 
 let controls = {};
@@ -311,672 +181,6 @@ function objectClick() {
   return;
 }
 
-class VoxelWorld {
-  constructor(options) {
-    this.id = options.id;
-    this.material = new THREE.MeshLambertMaterial({
-      map: texture,
-      side: THREE.DoubleSide,
-      opacity: 1,
-      transparent: true,
-    });
-    this.ux = options.ux;
-    this.uy = options.uy;
-    this.uz = options.uz;
-    this.x0 = options.x;
-    this.y0 = options.y;
-    this.z0 = options.z;
-    this.visible = true;
-    this.chunks = {};
-    this.absolute_positions = {};
-    this.chunkIdToMesh = {};
-    this.voxelcnt = 0;
-    this.clocks = new Set();
-  }
-
-  extract() {
-    const { id, ux, uy, uz, x0, y0, z0, chunks, voxelcnt } = this;
-    const chunkszip = {};
-    for (const [chunkId, chunk] of Object.entries(chunks)) {
-      var chunkzip = {};
-      for (const [voxelOffset, voxel] of Object.entries(chunk)) {
-        if (voxel != 0) {
-          chunkzip[voxelOffset] = voxel;
-        }
-      }
-      chunkszip[chunkId] = chunkzip;
-    }
-    return { id, ux, uy, uz, x0, y0, z0, voxelcnt, chunks: chunkszip };
-  }
-
-  load(loaded_world) {
-    const { ux, uy, uz, x0, y0, z0, voxelcnt, chunks } = loaded_world;
-    this.ux = ux;
-    this.uy = uy;
-    this.uz = uz;
-    this.x0 = x0;
-    this.y0 = y0;
-    this.z0 = z0;
-    this.voxelcnt = voxelcnt;
-    this.chunks = chunks;
-    for (let chunkId in chunks) {
-      var chunkIdinArray = chunkId.split(",");
-      this.updateVoxelGeometry(
-        Number(chunkIdinArray[0]) * chunkSize,
-        Number(chunkIdinArray[1]) * chunkSize,
-        Number(chunkIdinArray[2]) * chunkSize
-      );
-    }
-  }
-
-  computeVoxelOffset(x, y, z) {
-    const voxelX = THREE.MathUtils.euclideanModulo(x, chunkSize) | 0;
-    const voxelY = THREE.MathUtils.euclideanModulo(y, chunkSize) | 0;
-    const voxelZ = THREE.MathUtils.euclideanModulo(z, chunkSize) | 0;
-    return voxelY * chunkSliceSize + voxelZ * chunkSize + voxelX;
-  }
-  computeChunkId(x, y, z) {
-    const chunkX = Math.floor(x / chunkSize);
-    const chunkY = Math.floor(y / chunkSize);
-    const chunkZ = Math.floor(z / chunkSize);
-    return `${chunkX},${chunkY},${chunkZ}`;
-  }
-  addChunkForVoxel(x, y, z) {
-    const chunkId = this.computeChunkId(x, y, z);
-    let chunk = this.chunks[chunkId];
-    if (!chunk) {
-      chunk = new Uint8Array(chunkSize * chunkSize * chunkSize);
-      this.chunks[chunkId] = chunk;
-    }
-    return chunk;
-  }
-  getChunkForVoxel(x, y, z) {
-    return this.chunks[this.computeChunkId(x, y, z)];
-  }
-  setVoxel(x, y, z, v, addChunk = true) {
-    if (v == 0 && this.getVoxel(x, y, z) != 0) {
-      this.voxelcnt--;
-    } else if (v != 0 && this.getVoxel(x, y, z) == 0) {
-      this.voxelcnt++;
-    }
-    let chunk = this.getChunkForVoxel(x, y, z);
-    if (!chunk) {
-      if (!addChunk) {
-        return;
-      }
-      chunk = this.addChunkForVoxel(x, y, z);
-    }
-    const voxelOffset = this.computeVoxelOffset(x, y, z);
-    chunk[voxelOffset] = v;
-  }
-  getVoxel(x, y, z) {
-    const chunk = this.getChunkForVoxel(x, y, z);
-    if (!chunk) {
-      return 0;
-    }
-    const voxelOffset = this.computeVoxelOffset(x, y, z);
-    return chunk[voxelOffset];
-  }
-  generateGeometryDataForChunk(chunkX, chunkY, chunkZ) {
-    const positions = [];
-    const normals = [];
-    const uvs = [];
-    const indices = [];
-    const startX = chunkX * chunkSize;
-    const startY = chunkY * chunkSize;
-    const startZ = chunkZ * chunkSize;
-
-    for (let y = 0; y < chunkSize; ++y) {
-      const voxelY = startY + y;
-      for (let z = 0; z < chunkSize; ++z) {
-        const voxelZ = startZ + z;
-        for (let x = 0; x < chunkSize; ++x) {
-          const voxelX = startX + x;
-          const voxel = this.getVoxel(voxelX, voxelY, voxelZ);
-          if (voxel) {
-            // voxel 0 is sky (empty) so for UVs we start at 0
-            const uvVoxel = voxel - 1;
-            for (const { dir, corners, uvRow } of VoxelWorld.faces) {
-              const neighbor = this.getVoxel(
-                voxelX + dir[0],
-                voxelY + dir[1],
-                voxelZ + dir[2]
-              );
-              if (!neighbor) {
-                // this voxel has no neighbor in this direction so we need a face.
-                const ndx = positions.length / 3;
-                for (const { pos, uv } of corners) {
-                  positions.push(pos[0] + x, pos[1] + y, pos[2] + z);
-                  normals.push(...dir);
-                  uvs.push(
-                    ((uvVoxel + uv[0]) * tileSize) / tileTextureWidth,
-                    1 - ((uvRow + 1 - uv[1]) * tileSize) / tileTextureHeight
-                  );
-                }
-                indices.push(ndx, ndx + 1, ndx + 2, ndx + 2, ndx + 1, ndx + 3);
-              }
-            }
-          }
-        }
-      }
-    }
-
-    return {
-      positions,
-      normals,
-      uvs,
-      indices,
-    };
-  }
-
-  intersectRay(start, end) {
-    let dx = end.x - start.x;
-    let dy = end.y - start.y;
-    let dz = end.z - start.z;
-    const lenSq = dx * dx + dy * dy + dz * dz;
-    const len = Math.sqrt(lenSq);
-
-    dx /= len;
-    dy /= len;
-    dz /= len;
-
-    let t = 0.0;
-    let ix = Math.floor(start.x);
-    let iy = Math.floor(start.y);
-    let iz = Math.floor(start.z);
-
-    const stepX = dx > 0 ? 1 : -1;
-    const stepY = dy > 0 ? 1 : -1;
-    const stepZ = dz > 0 ? 1 : -1;
-
-    const txDelta = Math.abs(1 / dx);
-    const tyDelta = Math.abs(1 / dy);
-    const tzDelta = Math.abs(1 / dz);
-
-    const xDist = stepX > 0 ? ix + 1 - start.x : start.x - ix;
-    const yDist = stepY > 0 ? iy + 1 - start.y : start.y - iy;
-    const zDist = stepZ > 0 ? iz + 1 - start.z : start.z - iz;
-
-    // location of nearest voxel boundary, in units of t
-    let txMax = txDelta < Infinity ? txDelta * xDist : Infinity;
-    let tyMax = tyDelta < Infinity ? tyDelta * yDist : Infinity;
-    let tzMax = tzDelta < Infinity ? tzDelta * zDist : Infinity;
-
-    let steppedIndex = -1;
-
-    // main loop along raycast vector
-    while (t <= len) {
-      const voxel = this.getVoxel(ix, iy, iz);
-      if (voxel) {
-        return {
-          position: [start.x + t * dx, start.y + t * dy, start.z + t * dz],
-          normal: [
-            steppedIndex === 0 ? -stepX : 0,
-            steppedIndex === 1 ? -stepY : 0,
-            steppedIndex === 2 ? -stepZ : 0,
-          ],
-          voxel,
-        };
-      }
-
-      // advance t to next nearest voxel boundary
-      if (txMax < tyMax) {
-        if (txMax < tzMax) {
-          ix += stepX;
-          t = txMax;
-          txMax += txDelta;
-          steppedIndex = 0;
-        } else {
-          iz += stepZ;
-          t = tzMax;
-          tzMax += tzDelta;
-          steppedIndex = 2;
-        }
-      } else {
-        if (tyMax < tzMax) {
-          iy += stepY;
-          t = tyMax;
-          tyMax += tyDelta;
-          steppedIndex = 1;
-        } else {
-          iz += stepZ;
-          t = tzMax;
-          tzMax += tzDelta;
-          steppedIndex = 2;
-        }
-      }
-    }
-    return null;
-  }
-
-  updateChunkGeometry(x, y, z) {
-    const { material } = this;
-    const chunkX = Math.floor(x / chunkSize);
-    const chunkY = Math.floor(y / chunkSize);
-    const chunkZ = Math.floor(z / chunkSize);
-    const chunkId = `${chunkX},${chunkY},${chunkZ}`;
-    let mesh = this.chunkIdToMesh[chunkId];
-    const geometry = mesh ? mesh.geometry : new THREE.BufferGeometry();
-
-    const { positions, normals, uvs, indices } =
-      this.generateGeometryDataForChunk(chunkX, chunkY, chunkZ);
-    const positionNumComponents = 3;
-    geometry.setAttribute(
-      "position",
-      new THREE.BufferAttribute(
-        new Float32Array(positions),
-        positionNumComponents
-      )
-    );
-    const normalNumComponents = 3;
-    geometry.setAttribute(
-      "normal",
-      new THREE.BufferAttribute(new Float32Array(normals), normalNumComponents)
-    );
-    const uvNumComponents = 2;
-    geometry.setAttribute(
-      "uv",
-      new THREE.BufferAttribute(new Float32Array(uvs), uvNumComponents)
-    );
-    geometry.setIndex(indices);
-    geometry.computeBoundingSphere();
-
-    if (!mesh) {
-      mesh = new THREE.Mesh(geometry, material);
-      mesh.name = chunkId;
-      this.chunkIdToMesh[chunkId] = mesh;
-      scene.add(mesh);
-      mesh.position.set(0, 0, 0);
-      mesh.frustumCulled = false;
-      mesh.visible = this.visible;
-    }
-
-    this.absolute_positions[chunkId] = [];
-    for (let i = 0; i < positions.length / 3; i++) {
-      this.absolute_positions[chunkId].push({
-        x: chunkX * chunkSize + positions[3 * i] + this.x0,
-        y: chunkY * chunkSize + positions[3 * i + 1] + this.y0,
-        z: chunkZ * chunkSize + positions[3 * i + 2] + this.z0,
-      });
-    }
-
-    mesh.geometry.attributes.position.needsUpdate = true;
-  }
-
-  show() {
-    this.visible = true;
-    for (const chunkId in this.chunks) {
-      this.chunkIdToMesh[chunkId].visible = true;
-    }
-  }
-
-  setOpacity(opacity) {
-    this.material.opacity = opacity;
-    for (const chunkId in this.chunks) {
-      this.chunkIdToMesh[chunkId].material = this.material;
-    }
-  }
-
-  hide() {
-    this.visible = false;
-    for (const chunkId in this.chunks) {
-      this.chunkIdToMesh[chunkId].visible = false;
-    }
-  }
-
-  updateVoxelGeometry(x, y, z) {
-    const updatedChunkIds = {};
-    for (const offset of neighborOffsets) {
-      const ox = x + offset[0];
-      const oy = y + offset[1];
-      const oz = z + offset[2];
-      const chunkId = this.computeChunkId(ox, oy, oz);
-      if (!updatedChunkIds[chunkId]) {
-        updatedChunkIds[chunkId] = true;
-        if (!(chunkId in this.chunks)) {
-          this.addChunkForVoxel(ox, oy, oz);
-        }
-        this.updateChunkGeometry(ox, oy, oz);
-      }
-    }
-  }
-
-  updateRelativisticGeometry() {
-    const [ux, uy, uz] = [this.ux, this.uy, this.uz];
-    for (const chunkId in this.chunks) {
-      const positions =
-        this.chunkIdToMesh[chunkId].geometry.attributes.position.array;
-      var idx = 0;
-      this.absolute_positions[chunkId].forEach(function (point) {
-        if (editing) {
-          var [newx, newy, newz] = [
-            point.x - player.x,
-            point.y - player.y,
-            point.z - player.z,
-          ];
-          var time = 0;
-        } else {
-          if (usingRelativity == 0) {
-            var [newx, newy, newz] = [
-              point.x + ux * player.time - player.x,
-              point.y + uy * player.time - player.y,
-              point.z + uz * player.time - player.z,
-            ];
-            var time = player.time;
-          } else if (usingRelativity == 1) {
-            var [newx0, newy0, newz0] = doLengthContraction(
-              point.x,
-              point.y,
-              point.z,
-              ux,
-              uy,
-              uz
-            );
-            var [newx, newy, newz, time] = applyRelativity1(
-              newx0 - player.x,
-              newy0 - player.y,
-              newz0 - player.z,
-              player.vx,
-              player.vy,
-              player.vz,
-              ux,
-              uy,
-              uz,
-              player.time
-            );
-          } else {
-            var [newx0, newy0, newz0] = doLengthContraction(
-              point.x,
-              point.y,
-              point.z,
-              ux,
-              uy,
-              uz
-            );
-            var [newx, newy, newz, time] = applyRelativity2(
-              newx0 - player.x,
-              newy0 - player.y,
-              newz0 - player.z,
-              player.vx,
-              player.vy,
-              player.vz,
-              ux,
-              uy,
-              uz,
-              player.time
-            );
-          }
-        }
-        positions[idx++] = newx;
-        positions[idx++] = newy;
-        positions[idx++] = newz;
-      });
-
-      this.chunkIdToMesh[
-        chunkId
-      ].geometry.attributes.position.needsUpdate = true;
-    }
-  }
-
-  setVelocity({ newux, newuy, newuz }) {
-    this.ux = newux;
-    this.uy = newuy;
-    this.uz = newuz;
-
-    for (let clockId of this.clocks) {
-      clocks[clockId].setVelocity(newux, newuy, newuz);
-    }
-  }
-
-  findIntersection() {
-    const canvasX = 0.001;
-    const canvasY = -0.01;
-
-    const start = new THREE.Vector3();
-    const end = new THREE.Vector3();
-    start.set(
-      camera.position.x + player.x - this.x0,
-      camera.position.y + player.y - this.y0,
-      camera.position.z + player.z - this.z0
-    );
-    end.set(canvasX, canvasY, 1).unproject(camera);
-
-    const intersection = this.intersectRay(start, end);
-    return intersection;
-  }
-
-  placeVoxel(intersection) {
-    const voxelPos = intersection.position.map((v, ndx) => {
-      return v + intersection.normal[ndx] * 0.5;
-    });
-    this.setVoxel(...voxelPos, currentVoxel);
-    this.updateVoxelGeometry(...voxelPos);
-  }
-
-  deleteVoxel(intersection) {
-    const voxelPos = intersection.position.map((v, ndx) => {
-      return v + intersection.normal[ndx] * -0.5;
-    });
-    this.setVoxel(...voxelPos, 0);
-    this.updateVoxelGeometry(...voxelPos);
-  }
-
-  computeClockCoordinate(intersection) {
-    const clockPos = intersection.position.map((v, ndx) => {
-      return v + intersection.normal[ndx] * -0.5;
-    });
-    return [
-      clockPos[0] + this.x0,
-      clockPos[1] + this.y0,
-      clockPos[2] + this.z0,
-    ];
-  }
-}
-
-VoxelWorld.faces = [
-  {
-    // left
-    uvRow: 0,
-    dir: [-1, 0, 0],
-    corners: [
-      { pos: [0, 1, 0], uv: [0, 1] },
-      { pos: [0, 0, 0], uv: [0, 0] },
-      { pos: [0, 1, 1], uv: [1, 1] },
-      { pos: [0, 0, 1], uv: [1, 0] },
-    ],
-  },
-  {
-    // right
-    uvRow: 0,
-    dir: [1, 0, 0],
-    corners: [
-      { pos: [1, 1, 1], uv: [0, 1] },
-      { pos: [1, 0, 1], uv: [0, 0] },
-      { pos: [1, 1, 0], uv: [1, 1] },
-      { pos: [1, 0, 0], uv: [1, 0] },
-    ],
-  },
-  {
-    // bottom
-    uvRow: 1,
-    dir: [0, -1, 0],
-    corners: [
-      { pos: [1, 0, 1], uv: [1, 0] },
-      { pos: [0, 0, 1], uv: [0, 0] },
-      { pos: [1, 0, 0], uv: [1, 1] },
-      { pos: [0, 0, 0], uv: [0, 1] },
-    ],
-  },
-  {
-    // top
-    uvRow: 2,
-    dir: [0, 1, 0],
-    corners: [
-      { pos: [0, 1, 1], uv: [1, 1] },
-      { pos: [1, 1, 1], uv: [0, 1] },
-      { pos: [0, 1, 0], uv: [1, 0] },
-      { pos: [1, 1, 0], uv: [0, 0] },
-    ],
-  },
-  {
-    // back
-    uvRow: 0,
-    dir: [0, 0, -1],
-    corners: [
-      { pos: [1, 0, 0], uv: [0, 0] },
-      { pos: [0, 0, 0], uv: [1, 0] },
-      { pos: [1, 1, 0], uv: [0, 1] },
-      { pos: [0, 1, 0], uv: [1, 1] },
-    ],
-  },
-  {
-    // front
-    uvRow: 0,
-    dir: [0, 0, 1],
-    corners: [
-      { pos: [0, 0, 1], uv: [0, 0] },
-      { pos: [1, 0, 1], uv: [1, 0] },
-      { pos: [0, 1, 1], uv: [0, 1] },
-      { pos: [1, 1, 1], uv: [1, 1] },
-    ],
-  },
-];
-
-class Clock {
-  constructor(options) {
-    this.id = options.id;
-    this.name = options.name;
-    this.objectId = options.objectId;
-    this.labelelem = options.labelelem;
-    this.clockelem = options.clockelem;
-    this.clockcheckelem = options.clockcheckelem;
-    this.clockInstanceElem = options.clockInstanceElem;
-    this.ux = worlds[this.objectId].ux;
-    this.uy = worlds[this.objectId].uy;
-    this.uz = worlds[this.objectId].uz;
-    this.x = options.x;
-    this.y = options.y;
-    this.z = options.z;
-  }
-
-  extract() {
-    const { id, name, objectId, ux, uy, uz, x, y, z } = this;
-    return { id, name, objectId, ux, uy, uz, x, y, z };
-  }
-
-  update() {
-    const [ux, uy, uz] = [this.ux, this.uy, this.uz];
-    if (editing) {
-      var [newx, newy, newz] = [
-        this.x - player.x,
-        this.y - player.y,
-        this.z - player.z,
-      ];
-      var time = 0;
-    } else {
-      if (usingRelativity == 0) {
-        var [newx, newy, newz] = [
-          this.x + ux * player.time - player.x,
-          this.y + uy * player.time - player.y,
-          this.z + uz * player.time - player.z,
-        ];
-        var time = player.time;
-      } else if (usingRelativity == 1) {
-        var [newx0, newy0, newz0] = doLengthContraction(
-          this.x,
-          this.y,
-          this.z,
-          ux,
-          uy,
-          uz
-        );
-        var [newx, newy, newz, time] = applyRelativity1(
-          newx0 - player.x,
-          newy0 - player.y,
-          newz0 - player.z,
-          player.vx,
-          player.vy,
-          player.vz,
-          ux,
-          uy,
-          uz,
-          player.time
-        );
-      } else {
-        var [newx0, newy0, newz0] = doLengthContraction(
-          this.x,
-          this.y,
-          this.z,
-          ux,
-          uy,
-          uz
-        );
-        var [newx, newy, newz, time] = applyRelativity2(
-          newx0 - player.x,
-          newy0 - player.y,
-          newz0 - player.z,
-          player.vx,
-          player.vy,
-          player.vz,
-          ux,
-          uy,
-          uz,
-          player.time
-        );
-      }
-    }
-
-    const { name, labelelem, clockelem } = this;
-    clockelem.innerText = `${name} : ${time.toFixed(2)}s`;
-
-    const tempV = new THREE.Vector3(newx, newy, newz);
-    tempV.project(camera);
-    if (Math.abs(tempV.z) > 1) {
-      labelelem.style.display = "none";
-    } else {
-      labelelem.style.display = "block";
-      // convert the normalized position to CSS coordinates
-      if (
-        -0.95 < tempV.x &&
-        tempV.x < 0.95 &&
-        -0.93 < tempV.y &&
-        tempV.y < 0.93
-      ) {
-        const x =
-          (tempV.x * 0.5 + 0.5) * canvas.clientWidth -
-          labelelem.offsetWidth / 2;
-        const y =
-          (tempV.y * -0.5 + 0.5) * canvas.clientHeight -
-          labelelem.offsetHeight / 2;
-
-        // move the elem to that position
-        // labelelem.style.transform = `translate(${x}px,${y}px)`;
-        labelelem.style.left = `${x}px`;
-        labelelem.style.top = `${y}px`;
-
-        // 정렬을 위해 z-index 값을 설정합니다.
-        labelelem.style.zIndex = ((-tempV.z * 0.5 + 0.5) * 100000) | 0;
-      } else {
-        labelelem.style.display = "none";
-      }
-    }
-  }
-
-  setVelocity(ux, uy, uz) {
-    this.ux = ux;
-    this.uy = uy;
-    this.uz = uz;
-  }
-
-  showCheckbox() {
-    this.clockcheckelem.style.display = "inline";
-  }
-
-  hideCheckbox() {
-    this.clockcheckelem.style.display = "none";
-  }
-}
-
 const objectIdtoRadio = {};
 
 function createObject(id = worldNum) {
@@ -1036,31 +240,6 @@ function addWorldtoObjectRadio(val, objectId) {
   objectRadio.insertBefore(label, createObjectButton);
   radioInput.addEventListener("click", objectClick);
   objectIdtoRadio[objectId] = { radioInput, label };
-}
-
-function createWorld({ ux, uy, uz, x, y, z }, objectId) {
-  const newWorld = new VoxelWorld({
-    id: objectId,
-    ux,
-    uy,
-    uz,
-    x,
-    y,
-    z,
-  });
-  worlds[objectId] = newWorld;
-  currentObjectId = objectId;
-  newWorld.setVoxel(0, 0, 0, currentVoxel);
-  for (let chunkId in newWorld.chunks) {
-    var chunkIdinArray = chunkId.split(",");
-    newWorld.updateVoxelGeometry(
-      Number(chunkIdinArray[0]) * chunkSize,
-      Number(chunkIdinArray[1]) * chunkSize,
-      Number(chunkIdinArray[2]) * chunkSize
-    ); // 0,0,0 will generate
-  }
-  addWorldtoObjectRadio(worldNum + 1, objectId);
-  worldNum++;
 }
 
 function setWorldVelocity() {
@@ -1249,7 +428,6 @@ function extractData() {
   return data;
 }
 
-const worldNameForm = document.getElementById("worldnameform");
 const worldNameInput = document.getElementById("worldnameinput");
 worldNameInput.onfocus = lockControl;
 const worldNameSubmit = document.getElementById("worldnamesubmit");
@@ -1262,12 +440,6 @@ worldNameSubmit.onclick = function () {
   }
 };
 
-function prepareDownload() {
-  worldNameForm.style.display = "block";
-  downloadButton.blur();
-  worldNameInput.focus();
-}
-
 function download(worldName) {
   worldNameForm.style.display = "none";
   const content = JSON.stringify(extractData());
@@ -1278,17 +450,6 @@ function download(worldName) {
   a.href = URL.createObjectURL(file);
   a.download = fileName;
   a.click();
-}
-
-function openFile() {
-  loadFileButton.blur();
-  var input = document.createElement("input");
-  input.type = "file";
-  input.accept = "text/plain";
-  input.onchange = function (event) {
-    processFile(event.target.files[0]);
-  };
-  input.click();
 }
 
 function processFile(file) {
